abstract modal
